// src/app/models/character.model.ts
import { Location } from './location.model';
import { Origin } from './origin.model';

export interface Character {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: Origin;
  location: Location;
  image: string;
  episode: string[];
  url: string;
  created: string;
}